requirements.txt
kurigram
fastapi
uvicorn[standard]
uvloop
httptools
motor
redis[hiredis]
jinja2
python-dotenv
